refactor: refactors intervention info card logic and header component
Moves data extraction logic for intervention card to a shared utility,
improving code reuse and maintainability. Renames the header component
for clarity and updates references accordingly. Enhances organization
of related functionality.
