+++ src/features/projectsV2/ProjectDetails/components/OtherInterventionInfo.tsx
@@ -13,91 +13,8 @@ import TreeMapperBrand from './microComponents/TreeMapperBrand';
 import ImageSlider from './ImageSlider';
 import MobileInfoSwiper from '../../MobileInfoSwiper';
 import OtherInterventionMetadata from './microComponents/OtherInterventionMetadata';
-import OtherInterventionHeader from './microComponents/OtherInterventionHeader';
-
-interface MetaDataValue {
-  value: string;
-  label: string;
-}
-
-interface PublicMetaData {
-  [key: string]: string | MetaDataValue;
-}
-
-function isJsonString(str: string) {
-  try {
-    const parsed = JSON.parse(str);
-    return typeof parsed === 'object' && parsed !== null;
-  } catch (e) {
-    return false;
-  }
-}
-
-const createCardData = (interventionInfo: OtherInterventions | null) => {
-  // Initialize an array to store the cleaned key-value pairs
-  const cleanedData: { key: string; value: string }[] = [];
-
-  // Extract metadata from the interventionInfo object, if it exists
-  const parsedData = interventionInfo?.metadata;
-
-  // Check if `parsedData.public` exists, is an object, and is not an array
-  if (
-    parsedData?.public &&
-    typeof parsedData.public === 'object' &&
-    !Array.isArray(parsedData.public)
-  ) {
-    // Iterate over the entries of `parsedData.public` as key-value pairs
-    Object.entries(parsedData.public as PublicMetaData).forEach(
-      ([key, value]) => {
-        // Skip the entry if the key is 'isEntireSite' as it's used to show point location and no use to user
-        if (key !== 'isEntireSite') {
-          // If the value is a string, directly add it to cleanedData
-          if (typeof value === 'string') {
-            cleanedData.push({ value, key });
-          }
-          // If the value is an object with `value` and `label` properties
-          else if (
-            typeof value === 'object' &&
-            value !== null &&
-            'value' in value &&
-            'label' in value
-          ) {
-            // Check if the `value` property contains a valid JSON string
-            if (isJsonString(value.value)) {
-              try {
-                // Parse the JSON string
-                const parsedValue = JSON.parse(value.value);
-                // If the parsed value is an object with a `value` property, add it to cleanedData
-                if (
-                  parsedValue &&
-                  typeof parsedValue === 'object' &&
-                  'value' in parsedValue
-                ) {
-                  cleanedData.push({
-                    key: value.label, // Use the `label` property as the key
-                    value: parsedValue.value, // Use the parsed `value` property
-                  });
-                }
-              } catch (error) {
-                // Log an error if JSON parsing fails
-                console.error('Error parsing JSON:', error);
-              }
-            } else {
-              // If not a JSON string, add the `label` and `value` directly
-              cleanedData.push({
-                key: value.label,
-                value: value.value,
-              });
-            }
-          }
-        }
-      }
-    );
-  }
-
-  // Return the array of cleaned key-value pairs
-  return cleanedData;
-};
+import OtherInterventionInfoHeader from './microComponents/OtherInterventionInfoHeader';
+import { createCardData } from '../../../../utils/projectV2';
 
 interface Props {
   hoveredIntervention?: OtherInterventions | null;
@@ -151,7 +68,7 @@ const OtherInterventionInfo = ({
 
   const content = [
     <>
-      <OtherInterventionHeader
+      <OtherInterventionInfoHeader
         plHid={interventionInfo.hid}
         interventionType={interventionInfo.type}
         plantDate={interventionInfo.interventionStartDate}
